requirements.txt
python-telegram-bot==20.7
httpx==0.25.2
supabase==2.4.0
python-dotenv==1.0.0
